Group idea field definitions
